zincParser: Replace guts with hszinc module.
Use the implementation of the parser in hszinc, since I found the
standard one in pyhaystack is seriously deficient.  (e.g. decoding the
timezone "Melbourne" as "M" for instance).

Instead, we'll massage the Grid we get from hszinc into a dict.
Thankfully the row formats are the same.  We don't bother converting the
scalar values to JSON-grid format, since we'll only be converting them
back again; instead we return them as native Python or hszinc objects.