client.HaystackClient: Add mk_query for generating query strings.
Passing raw strings in URIs is a dangerous practice due to special
characters like & which have special meaning.  Correct method is to use
urllib to quote the components.

I'm assuming ordering is not important, so we'll pass these around in
keyword args for GET requests.  Still need to think of a good method for
POSTs.