client.ops.grid: Add retry logic.
This is necessary for NiagaraAX, which indicates a "logged out"
condition not by reporting error 401 like any sensible HTTP server
should, but by reporting error 200 and spitting back a HTML page.

The fact we didn't ask for HTML is the only clue that we've been booted
out.

We make a maximum of 3 attempts.  After that, we fail.