client.session: Add get_entities operation.
This operation is intended to retrieve entities by ID.  The objects
returned have the tags represented as attributes, for now just skeleton
objects.

Stubs are present to allow setting of tags / deleting of tags using the
standard Python operators, the results are then committed when the user
calls commit(), or may be reverted by calling revert().