client.entity.entity: Support entity instance invalidation.
It is plausable that between instantiation of an Entity instance, and
subsequent reads, that tags are added or removed that effectively change
the operations available for that entity.

e.g. a 'his' tag might be added or removed from a 'point' entity.

Since we can't change the type of a Python object, all we can do is
tell the caller to throw it away and fetch a new instance.  All
operations on the old entity should fail with an exception indicating
the stale nature of the instance.