client.ops.grid: Use 'body' not 'text'.
I'm not entirely certain that the encoding will be handled correctly by
the HTTP client implementation (it is in Requests, but what about
others?) so safer that we assume no, and do it ourselves.

'body' already contains decoded output and should be a unicode object in
Python 2.7 and a str object in Python 3.x.