client.http.sync: Execute callback as last action.
We should not call the callback until the very end.  Doing so here risks
the callback getting called twice in the event of an exception.

In fact, the callback should be called only ONCE, and we should swallow
the exceptions from it after possibly logging them to prevent crashing
any asynchronous IOLoop instances.