Fix PBKDF2 ImportError on Debian Wheezy (#46)
* Clean up whitespace changes.

- Remove carriage returns
- Remove trailing whitespace

* client.ops.grid: Drop `print`

We should *not* be printing stuff to `stdout` for several reasons:

1. who says the user is watching `stdout`?
2. who says the user isn't using `stdout` for something else?

Randomly printing stuff has the potential to corrupt `stdout` output of
any script that calls the grid operations, which is pretty much all
scripts.  We should not be polluting `stdout` with this data.

* util.scram: Add hashbang, encoding to file.

* util.scram: Trim trailing blank lines.

* util.scram: Split excessively long lines.

* util.scram: Use backports.pbkdf2 as fallback.

If we can't load `pbkdf2` from `hashlib`, use the backports version.