Synchronise VRT's Pyhaystack tree (#71)
* WC-847: client.widesky: Expect to see 404 messages.

* WC-847: client.ops.entity: Handle "not found" errors.

If we get back a `HaystackError` of the form `HNotFoundError: …`, then
the entity does not exist.

* WC-847: client.mixins.vendor.widesky.crud: Handle 200/400/404 error codes.

- 200: the "happy path"
- 400: bad request
- 404: not found

* WC-621: client.ops.his: Handle rng=slice object case.

MeterMaster passes in a `slice` object with two `datetime.datetime`
objects when it wishes to do a read operation.  It seems this construct
causes some grief with `pyhaystack` and `hszinc` which shouldn't be an
issue.

This fixes the issue by converting the slice object into a
comma-separated string: standard Project Haystack format.

* WC-1534: fix for creating entity without specifing 'id' tag

* WC-1534: add api server version check for uuid support

* VRT-1681: mkdeb.sh: Add Debian package build script.

* VRT-1681: Drop out-of-date Debian package files.

The script `mkdeb.sh` just needs `stdeb` and should remain up-to-date.

* WC-847: client.session: Wrap low-level function calls

This lets a subclass easily insert options into the `_get` or `_post`
request passed to the HTTP client, e.g. to expect certain HTTP error
codes.

* VRT-1681: client.widesky: Clean up duplication from merge.