Release 0 92 9 (#79)
* Syncing master with develop

Signed-off-by: Christian Tremblay <christian.tremblay@servisys.com>

* Quick fix to look for exception in scram login

* cleanup / fix setup.py, and add license (#69)

`re` and `requests` weren't used in this file, and weren't declared in
`setup_requires` causing some build issues. Removed these imports.

Removed mixed tabs & spaces.

Explicitly added the Apache License to list of Classifiers so this
project can be identified as Apache Licensed at pypi.org. This matches
the license file.

* Synchronise VRT's Pyhaystack tree (#71)

* WC-847: client.widesky: Expect to see 404 messages.

* WC-847: client.ops.entity: Handle "not found" errors.

If we get back a `HaystackError` of the form `HNotFoundError: …`, then
the entity does not exist.

* WC-847: client.mixins.vendor.widesky.crud: Handle 200/400/404 error codes.

- 200: the "happy path"
- 400: bad request
- 404: not found

* WC-621: client.ops.his: Handle rng=slice object case.

MeterMaster passes in a `slice` object with two `datetime.datetime`
objects when it wishes to do a read operation.  It seems this construct
causes some grief with `pyhaystack` and `hszinc` which shouldn't be an
issue.

This fixes the issue by converting the slice object into a
comma-separated string: standard Project Haystack format.

* WC-1534: fix for creating entity without specifing 'id' tag

* WC-1534: add api server version check for uuid support

* VRT-1681: mkdeb.sh: Add Debian package build script.

* VRT-1681: Drop out-of-date Debian package files.

The script `mkdeb.sh` just needs `stdeb` and should remain up-to-date.

* WC-847: client.session: Wrap low-level function calls

This lets a subclass easily insert options into the `_get` or `_post`
request passed to the HTTP client, e.g. to expect certain HTTP error
codes.

* VRT-1681: client.widesky: Clean up duplication from merge.

* Added a helper function to translate niagara ~2d, etc characters and make more readable strings

* Run black over the files.

* Update the company name of Widesky IoT platform (#78)

* cleanup / fix setup.py, and add license

`re` and `requests` weren't used in this file, and weren't declared in
`setup_requires` causing some build issues. Removed these imports.

Removed mixed tabs & spaces.

Explicitly added the Apache License to list of Classifiers so this
project can be identified as Apache Licensed at pypi.org. This matches
the license file.

* Update the company name of Widesky IoT platform

As of July 2019 Widesky.cloud is now the company responsible for developing Widesky.

* Added forgotten file

* Bumping version

* Trying to correct warning of deprecation for collections.abc (collections only won't work in Python 3.8)

* Formatting

* missing 2 in collection

* New encoding mixin for Niagara. Actually, added the unescaping function here.

* Will skip unescaping test for python 2... not supported

* Skip test

* Trying to set json by default for Niagara as zinc is unable to handle big requests.

* json by default for Niagara4

* Missing hszinc import

* format