# client.session: Use POST request to perform `hisRead`
SkySpark apparently do more than just "retrieve the data" when you do a
`hisRead`, and so have decided that it is not appropriate to allow `GET`
requests to perform a `hisRead`.
(Annoying, because it was handy to see the read range being fetched by a
client in HTTP server logs.)

# client.ops.grid: Allow pass-through of exclude_cookies.

SkySpark gets its knickers twisted if we "round-trip" cookies it sent
us.  So we need the ability to exclude the use of cookies in some cases.
# client.skyspark: Suppress cookies for SCRAM-based SkySpark
SkySpark really gets confused if you follow HTTP standards and hand its
cookies back to it.

# client.http.sync: Drop use of requests `Session`
This session object is responsible for doing things like
managing and expiring cookies, which normally is a desirable thing for
HTTP clients.
SkySpark gets horribly confused when you do this.  It will demand an
attestation key if it sees a cookie round-tripped by the HTTP client.
We actually round-trip things like cookies at a higher level, so it's
redundant for `requests` to do it for us, and plus it's harmful in this
case, we actually don't want to store the cookie in this situation.